+++ pkgs/jni/lib/src/jobject.dart
@@ -5,9 +5,10 @@
 import 'dart:ffi';
 
 import 'package:ffi/ffi.dart';
-import 'package:jni/internal_helpers_for_jnigen.dart';
 
+import 'accessors.dart';
 import 'jni.dart';
+import 'jreference.dart';
 import 'lang/jstring.dart';
 import 'types.dart';