+++ pkgs/jni/src/dartjni.h
@@ -4,9 +4,9 @@
 
 #pragma once
 
-// Note: include appropriate system jni.h as found by CMake, not third_party/jni.h.
 #include "include/dart_api_dl.h"
 
+// Note: include appropriate system jni.h as found by CMake, not third_party/jni.h.
 #include <jni.h>
 #include <stdint.h>
 #include <stdio.h>
@@ -392,9 +392,13 @@ JniResult PortProxy__newInstance(jobject binaryName,
 FFI_PLUGIN_EXPORT void resultFor(CallbackResult* result, jobject object);
 
 FFI_PLUGIN_EXPORT
-Dart_FinalizableHandle newFinalizableHandle(Dart_Handle object,
-                                            jobject reference,
-                                            jobjectRefType refType);
+Dart_FinalizableHandle newJObjectFinalizableHandle(Dart_Handle object,
+                                                   jobject reference,
+                                                   jobjectRefType refType);
+
+FFI_PLUGIN_EXPORT
+Dart_FinalizableHandle newBooleanFinalizableHandle(Dart_Handle object,
+                                                   bool* reference);
 
 FFI_PLUGIN_EXPORT
 void deleteFinalizableHandle(Dart_FinalizableHandle finalizableHandle,