+++ pkgs/jnigen/test/test_util/bindings_test_setup.dart
@@ -69,7 +69,6 @@ Future<void> bindingsTestSetup() async {
       kotlinTestJar,
     ]);
   }
-  Jni.initDLApi();
 }
 
 void bindingsTestTeardown() {