+++ pkgs/jni/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.8.0-wip
+## 0.8.0
 
 - **Breaking Change** ([#981](https://github.com/dart-lang/native/issues/981)):